Use uint32 for frame ID in F1 packet header
An unsigned 32-bit integer for the frame identifier in F1 packet headers
was incorrectly assumed to be an 8-bit unsigned integer. This previously
required a need to throw away 3 bytes before examining the packet body.
Examining frequent overflows in the frame identifier and incorrect
values for the player car index highlighted this issue.

+++ apps/apex/lib/f1/packet_header.ex
@@ -22,7 +22,7 @@ defmodule F1.PacketHeader do
     |> uint8(:packet_id)
     |> uint64(:session_uid)
     |> float32(:session_time)
-    |> uint8(:frame_identifier)
+    |> uint32(:frame_identifier)
     |> uint8(:player_car_index)
   end
 end
